requirements.txt
streamlit>=1.27.0
openai>=1.2.0
pillow>=9.4.0
python-dotenv>=1.0.0
